so much progress on everything

+++ FredsWorkmate/Database/Models/TrackedTime.cs
@@ -3,7 +3,7 @@
     public class TrackedTime : Model
     {
         public required Project Project { get; set; }
-        public required double Hours { get; set; }
+        public required decimal Hours { get; set; }
         public required DateOnly Date { get; set; }
     }
 }
